Use dumb-jump mode to jump across files
Oh-man ripgrep is so fast! :rocket: :rocket: :rocket: :rocket:

Although `dumb-jump` is language-agnostic meaning it does not understand the
language - cannot tell whether it is function usage or function call -
so shows all options found by `rg`. But for now this suffices my
needs! And most of all - it does not need any TAGS.